Updated Stock Trade records calculation

+++ src/main/kotlin/com/kotlin/assets/dto/ib/TradeRecord.kt
@@ -6,7 +6,7 @@ import java.time.LocalDate
 data class TradeRecord(
     val symbol: String,
     val tradeDate: LocalDate,
-    val quantity: BigDecimal,
+    val quantity: Int,
     val tradePrice: BigDecimal,
     val costBasis: BigDecimal,
     val ibCommission: BigDecimal,
